\name{addresses}
\alias{addresses}
\alias{airnow.pair.summ}
\alias{airnowdvs}
\alias{airnowtech}
\alias{census_00_centroids}
\alias{census_08}
\alias{contpm25.dva}
\alias{contpm25.dvd}
\alias{contpm25.pair.summ}
\alias{contpm25_all}
\alias{contpm25dvs}
\alias{latlongs}
\alias{o3.pair.summ}
\alias{o3dvs}
\alias{ozone.dvs}
\alias{ozone}
\alias{pm25.3day.pair.summ}
\alias{pm25.6day.pair.summ}
\alias{pm25.dvs}
\alias{pm25.dva}
\alias{pm25.dvd}
\alias{pm25avg}
\alias{pm25dvs}
\docType{data}
\title{
netasses data
}
\description{
  Datasets required for the netassess package
}
\usage{data("addresses")}
\format{
  Multiple data.frames used for the netassess package
}
\examples{
data(addresses)
## maybe str(addresses) ; plot(addresses) ...
}
\keyword{datasets}
